fix: pr review changes

+++ instrumentation/finagle-http-23.11/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/finaglehttp/v23_11/TwitterUtilCoreInstrumentationTest.java
@@ -112,7 +112,7 @@ class TwitterUtilCoreInstrumentationTest {
                           return 1;
                         })));
 
-    assertThat((int) Await.result(f, AWAIT)).isEqualTo(1);
+    assertThat(Await.result(f, AWAIT)).isEqualTo(1);
     assertParentAndChild("parent", "pool-child");
   }
